Allow for redundant timezone name and timezone offset.  Although, not ISO 8601 compliant, Java allows it: ISO_ZONED_DATE_TIME public static final DateTimeFormatter ISO_ZONED_DATE_TIME The ISO-like date-time formatter that formats or parses a date-time with offset and zone, such as '2011-12-03T10:15:30+01:00[Europe/Paris]'. This returns an immutable formatter capable of formatting and parsing a format that extends the ISO-8601 extended offset date-time format to add the time-zone. The section in square brackets is not part of the ISO-8601 standard. The format consists of:
The ISO_OFFSET_DATE_TIME
If the zone ID is not available or is a ZoneOffset then the format is complete.
An open square bracket '['.
The zone ID. This is not part of the ISO-8601 standard. Parsing is case sensitive.
A close square bracket ']'.
The returned formatter has a chronology of ISO set to ensure dates in other calendar systems are correctly converted. It has no override zone and uses the STRICT resolver style.