Streamling Converter to support "surrogate" pairs so that less base definition is required and more "expansion" of convertable pairs is offered.  An example - AtomicLong is a surrogate to Long (same for AtomicInteger to Integer and AtomicBoolean to Boolean, or StringBuilder and StringBuffer, to String).  So we only have to define the main class of the pair.  Then after those are loaded, a BFS expansion method is used to "blow out" the additional implied conversions, AtomicLong -> Long -> all things Long can be converted to.  This way, less static code is written and we obtain a larger set of conversions.
Separately, there was a bug in the Calendar to BigInteger/Long in that it was converting to microseconds.  This should have been milliseconds because these are older Java time classes.  For the newer Java temporal classes that support nanos, then for conversions to "integer" times (long/BigInteger) the units are epoch nanoseconds.